currency clean up
